Fixed import casing, organize server model imports
