+++ server/src/models/User.ts
@@ -1,7 +1,7 @@
-import mongoose, { Schema, Document } from "mongoose";
+import mongoose, { Document, Schema } from "mongoose";
 
 // TypeScript interface for the User model
-interface User extends Document {
+interface IUser extends Document {
 	email: string;
 	password: string;
 	mailingList: boolean;
@@ -9,11 +9,13 @@ interface User extends Document {
 }
 
 // Mongoose schema for the User model
-const UserSchema: Schema<User> = new Schema({
+const UserSchema: Schema<IUser> = new Schema({
 	email: { type: String, required: true },
 	password: { type: String, required: true },
 	mailingList: { type: Boolean, required: true },
 	role: { type: String, required: true },
 });
 
-export default mongoose.model<User>("User", UserSchema);
+const User = mongoose.model<IUser>("User", UserSchema);
+
+export default User;
