+++ server/src/models/index.ts
@@ -1,0 +1,2 @@
+export { default as Product } from "./Product";
+export { default as User } from "./User";
