--- server/src/models/userDetails.ts
@@ -1,6 +1,0 @@
-export interface IUser extends Document {
-	email: string;
-	password: string;
-	mailingList: boolean;
-	role: string;
-}
